move dataset registry to datasets module
