+++ nncore/learner/segmentationlearner.py
@@ -1,4 +1,3 @@
-from ..datasets import *
 from ..utils import *
 from ..schedulers import *
 from ..utils.utils import inverse_normalize_batch
